% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometric_models.R
\name{calculate_instrument_strength}
\alias{calculate_instrument_strength}
\title{Calculate Instrument Strength}
\usage{
calculate_instrument_strength(data)
}
\arguments{
\item{data}{Enhanced EPC data}
}
\value{
Data frame with instrument strength results
}
\description{
Calculate Instrument Strength
}
