% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_instruments.R
\name{create_alternative_sota_instruments}
\alias{create_alternative_sota_instruments}
\title{Create Alternative State-of-the-Art Instruments}
\usage{
create_alternative_sota_instruments(data)
}
\arguments{
\item{data}{Enhanced EPC data}
}
\value{
Data frame with alternative SOTA instruments
}
\description{
Create Alternative State-of-the-Art Instruments
}
