% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_entropy.R
\name{create_te_based_instruments}
\alias{create_te_based_instruments}
\title{Create Transfer Entropy-Based Instruments}
\usage{
create_te_based_instruments(data, te_results)
}
\arguments{
\item{data}{EPC data}

\item{te_results}{Transfer entropy analysis results}
}
\value{
List with enhanced data and network centralities
}
\description{
Create Transfer Entropy-Based Instruments
}
