% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{create_test_instruments}
\alias{create_test_instruments}
\title{Create test instruments for testing}
\usage{
create_test_instruments()
}
\value{
Data frame with test instruments
}
\description{
Create test instruments for testing
}
