% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_creation.R
\name{merge_epc_with_created_instruments}
\alias{merge_epc_with_created_instruments}
\title{Merge EPC Data with Created Instruments}
\usage{
merge_epc_with_created_instruments(epc_data, instruments)
}
\arguments{
\item{epc_data}{EPC data frame}

\item{instruments}{Instruments data frame}
}
\value{
Enhanced data frame with merged instruments
}
\description{
Merge EPC Data with Created Instruments
}
