% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_instrument_causal_pathways}
\alias{plot_instrument_causal_pathways}
\title{Create Instrument Causal Pathways Network}
\usage{
plot_instrument_causal_pathways(data, output_dir = NULL)
}
\arguments{
\item{data}{Enhanced EPC data}

\item{output_dir}{Directory to save plots (optional)}
}
\value{
ggplot object
}
\description{
Create Instrument Causal Pathways Network
}
