% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econometric_models.R
\name{run_comprehensive_epc_models}
\alias{run_comprehensive_epc_models}
\title{Run Comprehensive EPC Models}
\usage{
run_comprehensive_epc_models(data)
}
\arguments{
\item{data}{Enhanced EPC data with all instruments}
}
\value{
List of fitted models
}
\description{
Run Comprehensive EPC Models
}
