% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_epc_data}
\alias{sample_epc_data}
\title{Sample Environmental Phillips Curve Data}
\format{
A data frame with 155 rows and 9 variables:
\describe{
  \item{country}{Country name}
  \item{year}{Year (1991-2021)}
  \item{CO2_per_capita}{CO2 emissions per capita}
  \item{UR}{Total unemployment rate}
  \item{URF}{Female unemployment rate}
  \item{URM}{Male unemployment rate}
  \item{PCGDP}{Per capita GDP}
  \item{Trade}{Trade openness}
  \item{RES}{Renewable energy share}
}
}
\source{
Generated for package testing and demonstration
}
\usage{
sample_epc_data
}
\description{
A dataset containing Environmental Phillips Curve variables for 5 countries
from 1991 to 2021, used for testing and demonstration purposes.
}
\keyword{datasets}
