% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_creation.R
\name{create_composite_instruments}
\alias{create_composite_instruments}
\title{Create Composite Instruments using Factor Analysis}
\usage{
create_composite_instruments(instruments)
}
\arguments{
\item{instruments}{Data frame with individual instruments}
}
\value{
Enhanced data frame with composite instruments
}
\description{
Create Composite Instruments using Factor Analysis
}
